requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dateparser==1.2.0
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
notion-client==2.2.1
openai==1.14.3
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.7.1
tzlocal==5.2
yarl==1.9.4
